requirements.txt
streamlit
sqlalchemy
pymysql
pandas
matplotlib
pytest
python-dotenv
